+++ src/scrapers/movie/AdultDvdEmpire.cpp
@@ -147,7 +147,9 @@ void AdultDvdEmpire::onLoadFinished()
     if (reply->error() == QNetworkReply::NoError) {
         QString msg = QString::fromUtf8(reply->readAll());
         parseAndAssignInfos(msg, movie, reply->property("infosToLoad").value<Storage*>()->movieInfosToLoad());
+
     } else {
+        showNetworkError(*reply);
         qWarning() << "Network Error" << reply->errorString();
     }