+++ src/ui/movies/MovieSearch.h
@@ -26,6 +26,8 @@ public:
 public slots:
     int exec() override;
     int exec(QString searchString, ImdbId id, TmdbId tmdbId);
+
+public:
     static MovieSearch* instance(QWidget* parent = nullptr);
     QString scraperId();
     QString scraperMovieId();